fix tests, more split
